Set timezone in datetime inputs only on saving
