+++ src/editing/attributesComponent.js
@@ -1,6 +1,6 @@
 // The MIT License (MIT)
 //
-// Copyright (c) 2016-2021 Camptocamp SA
+// Copyright (c) 2016-2022 Camptocamp SA
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy of
 // this software and associated documentation files (the "Software"), to deal in
@@ -22,7 +22,6 @@
 import angular from 'angular';
 import {getUid as olUtilGetUid} from 'ol/util';
 import {listen} from 'ol/events';
-import DateFormatter from 'ngeo/misc/php-date-formatter';
 import ngeoMiscEventHelper from 'ngeo/misc/EventHelper';
 import ngeoMiscDatetimepickerComponent from 'ngeo/misc/datetimepickerComponent';
 import {ObjectEvent} from 'ol/Object';
@@ -175,13 +174,13 @@ Controller.prototype.$onInit = function () {
     throw new Error('Missing feature');
   }
   this.properties = this.feature.getProperties();
-  this.sanitize_();
 
   // Listen to the feature inner properties change and apply them to the form
   const uid = olUtilGetUid(this);
 
   // Set properties before starting to listen to changes to avoid multi AngularJS apply error.
   this.attributes.forEach((attribute) => {
+    this.sanitize_(attribute);
     if (
       attribute.type === 'boolean' &&
       (this.feature.getProperties()[attribute.name] === null ||
@@ -210,7 +209,8 @@ Controller.prototype.handleInputChange = function (name) {
     throw new Error('Missing feature');
   }
   this.updating_ = true;
-  this.sanitize_();
+  const attribute = this.attributes.find((attr) => attr.name === name);
+  this.sanitize_(attribute);
   const value = this.properties[name];
   this.feature.set(name, value);
   this.updating_ = false;
@@ -220,37 +220,20 @@ Controller.prototype.handleInputChange = function (name) {
  * Never keep a undefined values, use null.
  * On boolean, replace null by false.
  * On date, datetime and time replace empty string by null.
+ * @param {import('ngeo/format/Attribute.js').Attribute} attribute
  */
-Controller.prototype.sanitize_ = function () {
-  const dateFormatter = new DateFormatter();
-  this.attributes.forEach((attribute) => {
-    const value = this.properties[attribute.name];
-    if (value === undefined) {
-      this.properties[attribute.name] = null;
-    }
-    if (attribute.type === 'boolean' && value === null) {
-      this.properties[attribute.name] = false;
-    } else if (attribute.format) {
-      // Case of date, datetime or time.
-      if (value) {
-        console.assert(typeof value == 'string');
-        const formattedValue = dateFormatter.parseDate(value, attribute.format);
-        let jsonFormat = 'Y-m-d\\TH:i:s';
-        if (attribute.type === 'date') {
-          jsonFormat = 'Y-m-d';
-        } else if (attribute.type === 'time') {
-          jsonFormat = 'H:i:s';
-        } else if (attribute.type === 'datetime') {
-          // Time zone correction
-          formattedValue.setMinutes(formattedValue.getMinutes() + formattedValue.getTimezoneOffset());
-        }
-        this.properties[attribute.name] = dateFormatter.formatDate(formattedValue, jsonFormat);
-      } else {
-        // Shouldn't be set to an empty string
-        this.properties[attribute.name] = null;
-      }
-    }
-  });
+Controller.prototype.sanitize_ = function (attribute) {
+  const value = this.properties[attribute.name];
+  if (value === undefined) {
+    this.properties[attribute.name] = null;
+  }
+  if (attribute.type === 'boolean' && value === null) {
+    this.properties[attribute.name] = false;
+  } else if (attribute.format && !value) {
+    // Case of date, datetime or time.
+    // Shouldn't be set to an empty string
+    this.properties[attribute.name] = null;
+  }
 };
 
 /**
