+++ src/editing/editFeatureComponent.js
@@ -1,6 +1,6 @@
 // The MIT License (MIT)
 //
-// Copyright (c) 2016-2021 Camptocamp SA
+// Copyright (c) 2016-2022 Camptocamp SA
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy of
 // this software and associated documentation files (the "Software"), to deal in
@@ -744,6 +744,26 @@ Controller.prototype.save = function () {
   const id = this.featureId;
 
   this.pending = true;
+
+  // Correct datetime on save
+  const dateFormatter = new DateFormatter();
+  this.attributes.forEach((attribute) => {
+    const value = this.feature.get(attribute.name);
+    if (attribute.format && value) {
+      console.assert(typeof value == 'string');
+      const formattedValue = dateFormatter.parseDate(value, attribute.format);
+      let jsonFormat = 'Y-m-d\\TH:i:s';
+      if (attribute.type === 'date') {
+        jsonFormat = 'Y-m-d';
+      } else if (attribute.type === 'time') {
+        jsonFormat = 'H:i:s';
+      } else if (attribute.type === 'datetime') {
+        // Time zone correction
+        formattedValue.setMinutes(formattedValue.getMinutes() + formattedValue.getTimezoneOffset());
+      }
+      feature.set(attribute.name, dateFormatter.formatDate(formattedValue, jsonFormat));
+    }
+  });
 
   const promise = id
     ? this.gmfEditFeature_.updateFeature(this.editableNode_.id, feature)
